pyproject.toml
[project]
name = "framework-conv-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "javalang>=0.13.0",
    "mcp[cli]>=1.9.4",
    "openai>=1.93.0",
]
